requirements.txt
alembic==1.9.4
attrs==22.2.0
autopep8==2.0.2
beautifulsoup4==4.11.2
blinker==1.5
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
distlib==0.3.6
exceptiongroup==1.1.0
Flask==2.2.3
Flask-DebugToolbar==0.13.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.5
pycodestyle==2.10.0
pytest==7.2.2
python-dotenv==0.21.1
requests==2.28.2
soupsieve==2.4
SQLAlchemy==2.0.4
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
WTForms==3.0.1
